[#47016] Replace Users::TableCell by ViewComponent
https://community.openproject.org/work_packages/47016
